reoganize code structure
