+++ src/App.js
@@ -1,32 +1,17 @@
 import React from 'react';
-import { NavLink } from 'react-router-dom';
 
-import Router from './Router';
-// this Router is a router example, login has its own route part
-import { view as Login } from './login/';
-
-const Navigation = props => (
-  <nav>
-    <ul>
-      <li>
-        <NavLink to="/">Home</NavLink>
-      </li>
-      <li>
-        <NavLink to="/cart">Cart</NavLink>
-      </li>
-      <li>
-        <NavLink to="/login">Login</NavLink>
-      </li>
-    </ul>
-  </nav>
-);
+import { view as Home } from './features/food-listing';
+import { view as Navigation } from './features/navigation';
+import { view as Login } from './features/login';
+import { view as Cart } from './features/cart';
 
 function App() {
   return (
     <div className="page-container">
       <h1>Restaurant platform</h1>
       <Navigation />
-      <Router />
+      <Home />
+      <Cart />
       <Login />
     </div>
   );
