+++ src/features/cart/views/cartpage.js
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import Cart from '../features/cart';
+import Cart from './cart';
 
 export default function Cartpage(props) {
   return (
-    <div>
+    <div className="cart-page">
       <h2>Cartpage</h2>
       <Cart />
     </div>
