+++ src/features/food-listing/views/homepage.js
@@ -1,13 +1,13 @@
 import React from 'react';
 
-import FoodListing from '../features/food-listing';
+import FoodListing from './food-listing';
 
 // will get data from back-end
-import data from '../data/foods.json';
+import data from '../../../data/foods.json';
 
 export default function Homepage(props) {
   return (
-    <div>
+    <div className="food-listing-page">
       <h2>Homepage</h2>
       <FoodListing foods={data.foods} />
     </div>
