+++ src/features/login/actionTypes.js
@@ -1,0 +1,2 @@
+export const SUBMIT_LOGIN = 'LOGIN/SUBMIT';
+export const REGISTER_LOGIN = 'LOGIN/REGISTER';
