+++ src/features/login/index.js
@@ -1,0 +1,5 @@
+import * as actions from './actions';
+import reducer from './reducer';
+import view from './views/router';
+
+export { actions, reducer, view };
