+++ src/features/login/views/router.js
@@ -1,8 +1,8 @@
 import React from 'react';
+import { Switch, Route } from 'react-router-dom';
+
 import Login from './login';
 import Register from './register';
-import { Switch, Route } from 'react-router-dom';
-// import "./style.css";
 
 export default () => {
   return (
@@ -10,7 +10,6 @@ export default () => {
       <Switch>
         <Route exact path="/login" component={Login} />
         <Route exact path="/register" component={Register} />
-        {/* <Route path="/register" component={Register} /> */}
       </Switch>
     </div>
   );
