+++ src/features/navigation/index.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+
+const view = props => (
+  <nav>
+    <ul>
+      <li>
+        <NavLink to="/">Home</NavLink>
+      </li>
+      <li>
+        <NavLink to="/cart">Cart</NavLink>
+      </li>
+      <li>
+        <NavLink to="/login">Login</NavLink>
+      </li>
+    </ul>
+  </nav>
+);
+
+export { view };
