--- src/login/index.js
@@ -1,5 +1,0 @@
-import * as actions from "./actions";
-import reducer from "./reducer";
-import view from "./views/loginpage";
-
-export { actions, reducer, view };
